+++ src/features/registers-transfers/RegLayout.jsx
@@ -1,75 +1,11 @@
 import React from "react";
-import PremiumLearningShell from "../../shared/components/topics/PremiumLearningShell";
+import TopicLayout from "../../shared/components/topics/TopicLayout";
 import "./RegStyles.css";
+import { regPages, REG_TOPIC, REG_PATH_TO_SUBTOPIC_ID } from "./regConfig";
 
-const regPages = [
-  {
-    path: "/registers/intro",
-    label: "Registers",
-    description: "Register fundamentals, data storage roles, and system context.",
-  },
-  {
-    path: "/registers/counters",
-    label: "Counters",
-    description: "Counting circuits, sequences, and binary progression behavior.",
-  },
-  {
-    path: "/registers/sync-async",
-    label: "Synchronous / Asynchronous",
-    description: "Clocked coordination and the tradeoffs of async transitions.",
-  },
-  {
-    path: "/registers/shift-registers",
-    label: "Shift Registers",
-    description: "Move data bit by bit through serial and parallel structures.",
-  },
-  {
-    path: "/registers/serial-shift",
-    label: "Serial Shift Registers",
-    description: "Serial loading, shifting patterns, and timing of bit movement.",
-  },
-  {
-    path: "/registers/loading",
-    label: "Loading Registers",
-    description: "Control how data enters registers cleanly and predictably.",
-  },
-  {
-    path: "/registers/parallel",
-    label: "Parallel Registers",
-    description: "Parallel transfer techniques for wider, faster data movement.",
-  },
-  {
-    path: "/registers/ripple-counters",
-    label: "Ripple Counters",
-    description: "Asynchronous counter propagation and cumulative delay effects.",
-  },
-  {
-    path: "/registers/sync-binary-counters",
-    label: "Synchronous Binary Counters",
-    description: "Tighter clocked counter design with coordinated state changes.",
-  },
-];
-
-const PATH_TO_SUBTOPIC_ID = {
-  "/registers/intro": "registers",
-  "/registers/counters": "counters",
-  "/registers/sync-async": "sync-async",
-  "/registers/shift-registers": "shift-registers",
-  "/registers/serial-shift": "serial-shift",
-  "/registers/loading": "loading",
-  "/registers/parallel": "parallel",
-  "/registers/ripple-counters": "ripple-counters",
-  "/registers/sync-binary-counters": "sync-binary-counters",
-};
-
-const REG_TOPIC = {
-  id: "registers-and-register-transfers",
-  title: "REGISTERS & REGISTER TRANSFERS",
-  links: Object.values(PATH_TO_SUBTOPIC_ID).map((id) => ({ id })),
-};
 
 const RegLayout = ({ children, title, subtitle }) => (
-  <PremiumLearningShell
+  <TopicLayout
     title={title}
     subtitle={subtitle}
     pages={regPages}
@@ -81,11 +17,11 @@ const RegLayout = ({ children, title, subtitle }) => (
     rootClassName="reg-layout"
     tracking={{
       topic: REG_TOPIC,
-      pathToSubtopicId: PATH_TO_SUBTOPIC_ID,
+      pathToSubtopicId: REG_PATH_TO_SUBTOPIC_ID,
     }}
   >
     {children}
-  </PremiumLearningShell>
+  </TopicLayout>
 );
 
 export default RegLayout;
